[!!!][TASK] Migrate lowlevel deleted records command to SymfonyConsole
The lowlevel cleaner command for permanently delete records
in the database that have been previously marked as "deleted=1"
in the database is migrated to a Symfony Console command,
reducing the complexity and enhances the readability of the function.

Call it like this:
typo3/sysext/core/bin/typo3 cleanup:deletedrecords --dry-run -vv --pid=49 --depth=4

You can also use "-p" instead of "--pid", or "-d" instead of "--depth".

Resolves: #78417
Releases: master
Change-Id: I79fb2292d96c38c1406896cbe9d0bac6494d1fa9
Reviewed-on: https://review.typo3.org/50358
Tested-by: TYPO3com <no-reply@typo3.com>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>