[FEATURE] Introduce command to clean up processed files
This adds a new symfony command to `EXT:lowlevel`. The
command deletes processed files which are not referenced
anymore or in case their original file got removed.

Furthermore all `sys_file_processedfile` records referencing
on non-existing files are deleted, too. The command is also
capable of a "dry-run" option to just display the files and
records which would be deleted.

Resolves: #45039
Releases: main
Change-Id: Ib173331922d698584d970c2648751fd5d1937260
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/58721
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>