[TASK] Use Workspace API for checking TCA versioning support
The API method BackendUtility::isTableWorkspaceEnabled($table)
should be used when dealing with a backend context to determine
if a table has versioning support enabled.

This functionality wasn't used consistently throughout TYPO3
Backend / CLI code and should be used in order to determine.

For PageRepository a separate protected method is introduced
to avoid a dependency on BackendUtility.

This patch is a precursor for streamlining the workspaces API further.

Resolves: #89108
Releases: master
Change-Id: Id63e04a5faa1520459d23fb5467ebfb523d9d958
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/61639
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Susanne Moog <look@susi.dev>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Susanne Moog <look@susi.dev>