[TASK] Remove unused code related to internal_type=file
The internal_type=file was removed in TYPO3 v10, but the Backend Module
"DB Check" (lowlevel) still contained lots of parts that are not relevant
anymore.

This patch cleans up the corresponding parts and simplifies various other
places due to the lack of checks of internal_type=file.

Resolves: #90816
Releases: master
Change-Id: I742fc1ea197837c477415368a674ad8831e10f31
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/63872
Tested-by: TYPO3com <noreply@typo3.com>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>