[TASK] Align configuration module search bar
As a follow up of #94042, the search bar in
the configuration module is now also aligned
with the rest of the backend. The regular
expression checkbox is therefore moved into
a dropdown, according to the recordlist.

Since browsers nowadays add their own clearable
functionality to type=search fields, it's sufficient
to just register an event for it. Especially because
initialization of our custom clearable functionality
would otherwise lead to duplicated clear icons.

Resolves: #94101
Releated: #94042
Releases: master
Change-Id: I019d1de724d87a9ab074aec062c39295d9e41670
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69023
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Daniel Goerz <daniel.goerz@posteo.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Daniel Goerz <daniel.goerz@posteo.de>