[BUGFIX] Avoid PHP TypeError in EXT:lowlevel
EXT:lowlevel gathers some statistics, e.g. deleted
records in general. However, by looking at the code,
recStats[deleted] is grouped by TCA types, but 4 lines
later it is treated as a numeric integer.

This never worked, and actually throws a notice (and in PHP8 a
TypeError).

This code can be removed, as it never worked properly.

Resolves: #94176
Releases: master, 10.4
Change-Id: I7f6879f6c9f3b250b0f1571db62ccd5be60392a1
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/69219
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>