[BUGFIX] Drop input-group-btn usages if possible
The class input-group-btn was removed in bootstrap 4.6, we are now
dropping leftover usages. There is still a leftover in the form
framework rendering we cannot remove that easily. We are keeping
it as is for now since it would require refactoring of the form
javascript.

The other place in combination with typo3-mfa-totp-url-info-button
is handled in a dedicated patch.

Resolves: #100463
Releases: main
Change-Id: I5dbed36ff2b4c1e3e652523c488ed12284f20379
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78475
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>