[TASK] Drop special styling for popover, tooltips, h1, h2 and h3
There is not need for a special styling here, the components
are fine as is. We are now preferring the defaults over the
custom adjustments.

In rare cases the alignment needed to be adjusted for the
not missing default margin top of h2 and h3. These are
now hotfixed with manual adjustments until we managed
to solve the root problem with custom grids not providing
enough space for the next element.

Resolves: #100516
Releases: main
Change-Id: Ib7354524a3487247cbb27064e50c38d2099b886e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78523
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Markus Klein <markus.klein@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>