[TASK] Use local variable in `CleanUpLocalProcessedFilesCommand`
With recent changes in phpstan/phpstan, it seems the possible return
value of methods is not set in stone anymore, allowing detection
of possible issues in the code.

This change uses a local method variable for the method return
value of `isVerbose()` and uses this instead of several method calls.

Resolves: #100941
Releases: main, 12.4
Change-Id: If4a91d544e43b03279711dd8d5fbbb12c349566f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79127
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: core-ci <typo3@b13.com>