[TASK] Make "Show sys_log entries" non-schedulable
The command "Show entries from sys_log database" (syslog:list)
should not be schedulable because it is an output only
command and no output will be displayed in scheduler.

Also, no output is displayed if a scheduler task is
created for it and this is run from the shell.

The command is now no longer schedulable and is not
available as scheduler task.

Resolves: #101012
Releases: main, 12.4
Change-Id: I92648d1a81f56b49e7d9ad48dc82f9b0df0f6b39
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79398
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>