[BUGFIX] Replace method returnWebmounts() with getWebmounts()
The method BackendUserAuthentication::returnWebmounts() returns
an array of strings. This triggers a type error when calling
DatabaseIntegrityController::getTreeList() since this method
now requires the mountpoint as an integer.

In order to keep backwards compatibility a new method
BackendUserAuthentication::getWebmounts() has been introduced,
which returns an array of integers.

Existing calls to returnWebmounts() have been replaced by calls
to getWebmounts() and fixes by applying array_map(intval(...)
on the result is now superfluous.

A follow up issue has been created #104607 to deprecate method
returnWebmounts() and to investigate if	succesive calls	to
array_unique can be done in getWebmounts() and removed.

Resolves: #104591
Related: #104607
Releases: main, 12.4
Change-Id: I78b2ac3ca891b4144c3b68fd44afb645b875fa6f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/85614
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>