[TASK] Consolidate EXT:lowlevel labels
The record statistics submodule was moved to EXT:reports,
so the labels are now moved there, too.

Labels defined in EXT:core, which are only used in EXT:lowlevel,
are now moved into the scope of EXT:lowlevel.

Unused labels are deprecated, and adjusted labels utilize the
new domain syntax.

Resolves: #107923
Releases: main
Change-Id: Ife7b4a9c03a1ae12f8d55af57ef62407bdf59cee
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91349
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Garvin Hicking <garvin@hick.ing>
Tested-by: Garvin Hicking <garvin@hick.ing>