[BUGFIX] Prevent double initialization of clearable in datepicker
When using flatpickr with altInput, the clearable functionality was
being initialized twice: once on the original input field and once
on the flatpickr altInput. This resulted in duplicate clear buttons
being rendered.

This change removes the global markup trigger from places where they
have not been removed in the past (#107738).

Resolves: #108146
Related: #107738
Releases: main, 13.4
Change-Id: I8178a39ae70610e7f2ccef57ff641e7d6899c940
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/91703
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Kott <benjamin.kott@outlook.com>
Reviewed-by: Benjamin Kott <benjamin.kott@outlook.com>